requirements.txt
# django
django==5.1.1
django-admin-extras==1.0.2
django-cleanup==9.0.0
django-timezone-field==7.0.0
django-tinymce==4.1.0
django-cors-headers==4.4.0
django-allauth==65.0.2
django-widget-tweaks==1.5.0
django-compressor==4.5.1
django-bootstrap5==24.3
django-resized==1.0.2
django-sass-processor==1.4.1
django-cache-cleaner==0.1.0
django-simple-captcha==0.6.0
django-nonrelated-inlines==0.2

# general
requests==2.32.3
libsass==0.23.0

# local
tzdata==2024.2

# image
sorl-thumbnail==12.11.0
pillow==10.4.0

# rest framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1

# payment
stripe==10.12.0

# test
coverage==7.6.1

# formatter
black==24.8.0
babel==2.16.0
